# WebClient/HttpClients/Generated/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/matching/substitutes": {
      "post": {
        "tags": [
          "matching-controller"
        ],
        "operationId": "substitutesMatchRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/matching/gigs": {
      "post": {
        "tags": [
          "matching-controller"
        ],
        "operationId": "gigsMatchRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/job-confirmation/": {
      "post": {
        "tags": [
          "job-confirmation-controller"
        ],
        "operationId": "createJobConfirmation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobConfirmationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JobConfirmationDTO"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "job-confirmation-controller"
        ],
        "operationId": "answerJobConfirmation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerJobConfirmationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JobConfirmationDTO"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "sendMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDTO"
                }
              }
            }
          }
        }
      }
    },
    "/chat/": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "createChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BasicChatDTO"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JwtResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/test/yo": {
      "get": {
        "tags": [
          "auth-test-controller"
        ],
        "operationId": "test",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/matching/substitutes/{currentUserId}": {
      "get": {
        "tags": [
          "matching-controller"
        ],
        "operationId": "getSubstitutes",
        "parameters": [
          {
            "name": "currentUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubstituteMatchingDTOs"
                }
              }
            }
          }
        }
      }
    },
    "/matching/gigs/{currentUserId}": {
      "get": {
        "tags": [
          "matching-controller"
        ],
        "operationId": "getGigs",
        "parameters": [
          {
            "name": "currentUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GigMatchingDTOs"
                }
              }
            }
          }
        }
      }
    },
    "/chat/user/{userId}": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "getChatOverviewByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOverviewDTO"
                }
              }
            }
          }
        }
      }
    },
    "/chat/history/{id}": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "getChatHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/chat/gigs/{employerId}": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "getEmployerGigs",
        "parameters": [
          {
            "name": "employerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerGigsDTO"
                }
              }
            }
          }
        }
      }
    },
    "/chat/gig/{gigId}": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "getChatOverviewByGig",
        "parameters": [
          {
            "name": "gigId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOverviewDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MatchRequestDTO": {
        "type": "object",
        "properties": {
          "currentUser": {
            "type": "integer",
            "format": "int32"
          },
          "matchId": {
            "type": "integer",
            "format": "int32"
          },
          "wantsToMatch": {
            "type": "boolean"
          }
        }
      },
      "CreateJobConfirmationDTO": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "substituteId": {
            "type": "integer",
            "format": "int32"
          },
          "employerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "JobConfirmationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "substituteId": {
            "type": "integer",
            "format": "int32"
          },
          "employerId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNANSWERED",
              "ACCEPTED",
              "DECLINED"
            ]
          },
          "offeredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendMessageDTO": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "MessageDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateChatDTO": {
        "type": "object",
        "properties": {
          "substituteId": {
            "type": "integer",
            "format": "int32"
          },
          "employerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BasicChatDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "substituteId": {
            "type": "integer",
            "format": "int32"
          },
          "employerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LoginRequestDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "JwtResponseDTO": {
        "type": "object",
        "properties": {
          "jwtToken": {
            "type": "string"
          }
        }
      },
      "AnswerJobConfirmationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNANSWERED",
              "ACCEPTED",
              "DECLINED"
            ]
          }
        }
      },
      "SubstituteMatchingDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SubstituteMatchingDTOs": {
        "type": "object",
        "properties": {
          "possibleMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubstituteMatchingDTO"
            }
          }
        }
      },
      "GigMatchingDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "GigMatchingDTOs": {
        "type": "object",
        "properties": {
          "possibleMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GigMatchingDTO"
            }
          }
        }
      },
      "ChatOverviewDTO": {
        "type": "object",
        "properties": {
          "chats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicChatDTO"
            }
          }
        }
      },
      "ChatEmployerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "ChatHistoryDTO": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDTO"
            }
          },
          "jobConfirmation": {
            "$ref": "#/components/schemas/JobConfirmationDTO"
          },
          "employer": {
            "$ref": "#/components/schemas/ChatEmployerDTO"
          },
          "substitute": {
            "$ref": "#/components/schemas/ChatSubstituteDTO"
          }
        }
      },
      "ChatSubstituteDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "EmployerGigsDTO": {
        "type": "object",
        "properties": {
          "gigs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GigDTO"
            }
          }
        }
      },
      "GigDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}